replace sync manager with app
