--- wrappers/src/sync_manager_cs.hpp
@@ -1,38 +1,0 @@
-////////////////////////////////////////////////////////////////////////////
-//
-// Copyright 2017 Realm Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-////////////////////////////////////////////////////////////////////////////
-
-#pragma once
-
-#include "realm_export_decls.hpp"
-#include "sync/sync_config.hpp"
-
-using namespace realm;
-
-namespace realm {
-    class SyncUser;
-
-    struct SyncConfiguration
-    {
-        std::shared_ptr<SyncUser>* user;
-
-        uint16_t* url;
-        size_t url_len;
-
-        realm::ClientResyncMode client_resync_mode;
-    };
-}
