+++ wrappers/src/sync_user_cs.cpp
@@ -32,38 +32,6 @@ using SharedSyncSession = std::shared_ptr<SyncSession>;
 
 extern "C" {
 
-REALM_EXPORT SharedSyncUser* realm_get_current_sync_user(NativeException::Marshallable& ex)
-{
-    return handle_errors(ex, [&]() -> SharedSyncUser* {
-        auto ptr = SyncManager::shared().get_current_user();
-        if (ptr == nullptr) {
-            return nullptr;
-        }
-
-        return new SharedSyncUser(std::move(ptr));
-    });
-}
-
-REALM_EXPORT size_t realm_get_logged_in_users(SharedSyncUser** buffer, size_t buffer_length, NativeException::Marshallable& ex)
-{
-    return handle_errors(ex, [&]() -> size_t {
-        auto users = SyncManager::shared().all_users();
-        if (users.size() > buffer_length) {
-            return users.size();
-        }
-
-        if (users.size() <= 0) {
-            return 0;
-        }
-
-        for (size_t i = 0; i < users.size(); i++) {
-            buffer[i] = new SharedSyncUser(users.at(i));
-        }
-
-        return users.size();
-    });
-}
-
 REALM_EXPORT void realm_syncuser_log_out(SharedSyncUser& user, NativeException::Marshallable& ex)
 {
     handle_errors(ex, [&] {
